requirements.txt
beautifulsoup4~=4.8.2
click~=7.1.2
ipyplot~=1.1.0
lxml~=4.5.2
matplotlib~=3.3.2
numpy~=1.20.3
odfpy
opencv-python~=4.5.3.56
pandas~=1.1.3
pillow~=8.3.2
pymongo
pytest~=6.1.0
PyYAML~=5.3.1
requests~=2.23.0
sacred~=0.8.1
scikit-learn~=0.24.2
scipy~=1.4.1
seaborn~=0.11.2
snakemake~=5.17.0
statsmodels~=0.12.0
tabulate
torch>=1.6.0
torchvision~=0.11.1+cu113
tqdm~=4.46.0
umap-learn